Re-use existing tests for source gen
