+++ src/libraries/System.Text.Json/tests/System.Text.Json.Tests/Serialization/MetadataTests/DefaultJsonTypeInfoResolverTests.JsonPropertyInfo.cs
@@ -1690,338 +1690,6 @@ namespace System.Text.Json.Serialization.Tests
             Assert.Equal("NULL", value.Value);
         }
 
-        [Fact]
-        public static void RequiredAttributesGetDetectedAndFailDeserializationWhenValuesNotPresent()
-        {
-            JsonTypeInfo typeInfo = JsonSerializerOptions.Default.GetTypeInfo(typeof(ClassWithRequiredCustomAttributes));
-            Assert.Equal(3, typeInfo.Properties.Count);
-
-            Assert.Equal(nameof(ClassWithRequiredCustomAttributes.NonRequired), typeInfo.Properties[0].Name);
-            Assert.False(typeInfo.Properties[0].IsRequired);
-
-            Assert.Equal(nameof(ClassWithRequiredCustomAttributes.RequiredA), typeInfo.Properties[1].Name);
-            Assert.True(typeInfo.Properties[1].IsRequired);
-
-            Assert.Equal(nameof(ClassWithRequiredCustomAttributes.RequiredB), typeInfo.Properties[2].Name);
-            Assert.True(typeInfo.Properties[2].IsRequired);
-
-            ClassWithRequiredCustomAttributes obj = new();
-            string json = """{"NonRequired":null,"RequiredA":null,"RequiredB":null}""";
-            Assert.Equal(json, JsonSerializer.Serialize(obj));
-
-            var deserialized = JsonSerializer.Deserialize<ClassWithRequiredCustomAttributes>(json);
-            Assert.Null(deserialized.NonRequired);
-            Assert.Null(deserialized.RequiredA);
-            Assert.Null(deserialized.RequiredB);
-
-            JsonException exception = Assert.Throws<JsonException>(() => JsonSerializer.Deserialize<ClassWithRequiredCustomAttributes>("{}"));
-            Assert.DoesNotContain(nameof(ClassWithRequiredCustomAttributes.NonRequired), exception.Message);
-            Assert.Contains(nameof(ClassWithRequiredCustomAttributes.RequiredA), exception.Message);
-            Assert.Contains(nameof(ClassWithRequiredCustomAttributes.RequiredB), exception.Message);
-
-            exception = Assert.Throws<JsonException>(() => JsonSerializer.Deserialize<ClassWithRequiredCustomAttributes>("""{"NonRequired":"foo"}"""));
-            Assert.DoesNotContain(nameof(ClassWithRequiredCustomAttributes.NonRequired), exception.Message);
-            Assert.Contains(nameof(ClassWithRequiredCustomAttributes.RequiredA), exception.Message);
-            Assert.Contains(nameof(ClassWithRequiredCustomAttributes.RequiredB), exception.Message);
-
-
-            exception = Assert.Throws<JsonException>(() => JsonSerializer.Deserialize<ClassWithRequiredCustomAttributes>("""{"NonRequired":"foo", "RequiredB":"bar"}"""));
-            Assert.DoesNotContain(nameof(ClassWithRequiredCustomAttributes.NonRequired), exception.Message);
-            Assert.Contains(nameof(ClassWithRequiredCustomAttributes.RequiredA), exception.Message);
-            Assert.DoesNotContain(nameof(ClassWithRequiredCustomAttributes.RequiredB), exception.Message);
-
-            exception = Assert.Throws<JsonException>(() => JsonSerializer.Deserialize<ClassWithRequiredCustomAttributes>("""{"NonRequired":"foo", "RequiredA":null}"""));
-            Assert.DoesNotContain(nameof(ClassWithRequiredCustomAttributes.NonRequired), exception.Message);
-            Assert.DoesNotContain(nameof(ClassWithRequiredCustomAttributes.RequiredA), exception.Message);
-            Assert.Contains(nameof(ClassWithRequiredCustomAttributes.RequiredB), exception.Message);
-        }
-
-        [Fact]
-        public static void RequiredMemberCanBeModifiedToNonRequired()
-        {
-            JsonSerializerOptions options = new()
-            {
-                TypeInfoResolver = new DefaultJsonTypeInfoResolver()
-                {
-                    Modifiers =
-                    {
-                        ti =>
-                        {
-                            if (ti.Type == typeof(ClassWithRequiredCustomAttributes))
-                            {
-                                JsonPropertyInfo prop = ti.Properties[1];
-                                Assert.Equal(nameof(ClassWithRequiredCustomAttributes.RequiredA), prop.Name);
-                                Assert.True(prop.IsRequired);
-                                prop.IsRequired = false;
-                                Assert.False(prop.IsRequired);
-                            }
-                        }
-                    }
-                }
-            };
-
-            ClassWithRequiredCustomAttributes obj = new();
-            string json = """{"NonRequired":null,"RequiredA":null,"RequiredB":null}""";
-            Assert.Equal(json, JsonSerializer.Serialize(obj, options));
-
-            var deserialized = JsonSerializer.Deserialize<ClassWithRequiredCustomAttributes>(json, options);
-            Assert.Null(deserialized.NonRequired);
-            Assert.Null(deserialized.RequiredA);
-            Assert.Null(deserialized.RequiredB);
-
-            deserialized = JsonSerializer.Deserialize<ClassWithRequiredCustomAttributes>("""{"NonRequired":"foo", "RequiredB":"bar"}""", options);
-            Assert.Equal("foo", deserialized.NonRequired);
-            Assert.Null(deserialized.RequiredA);
-            Assert.Equal("bar", deserialized.RequiredB);
-
-            JsonException exception = Assert.Throws<JsonException>(() => JsonSerializer.Deserialize<ClassWithRequiredCustomAttributes>("{}", options));
-            Assert.DoesNotContain(nameof(ClassWithRequiredCustomAttributes.NonRequired), exception.Message);
-            Assert.DoesNotContain(nameof(ClassWithRequiredCustomAttributes.RequiredA), exception.Message);
-            Assert.Contains(nameof(ClassWithRequiredCustomAttributes.RequiredB), exception.Message);
-
-            exception = Assert.Throws<JsonException>(() => JsonSerializer.Deserialize<ClassWithRequiredCustomAttributes>("""{"NonRequired":"foo"}""", options));
-            Assert.DoesNotContain(nameof(ClassWithRequiredCustomAttributes.NonRequired), exception.Message);
-            Assert.DoesNotContain(nameof(ClassWithRequiredCustomAttributes.RequiredA), exception.Message);
-            Assert.Contains(nameof(ClassWithRequiredCustomAttributes.RequiredB), exception.Message);
-
-            exception = Assert.Throws<JsonException>(() => JsonSerializer.Deserialize<ClassWithRequiredCustomAttributes>("""{"NonRequired":"foo", "RequiredA":null}""", options));
-            Assert.DoesNotContain(nameof(ClassWithRequiredCustomAttributes.NonRequired), exception.Message);
-            Assert.DoesNotContain(nameof(ClassWithRequiredCustomAttributes.RequiredA), exception.Message);
-            Assert.Contains(nameof(ClassWithRequiredCustomAttributes.RequiredB), exception.Message);
-        }
-
-        [Fact]
-        public static void NonRequiredMemberCanBeModifiedToRequired()
-        {
-            JsonSerializerOptions options = new()
-            {
-                TypeInfoResolver = new DefaultJsonTypeInfoResolver()
-                {
-                    Modifiers =
-                    {
-                        ti =>
-                        {
-                            if (ti.Type == typeof(ClassWithRequiredCustomAttributes))
-                            {
-                                JsonPropertyInfo prop = ti.Properties[0];
-                                Assert.Equal(nameof(ClassWithRequiredCustomAttributes.NonRequired), prop.Name);
-                                Assert.False(prop.IsRequired);
-                                prop.IsRequired = true;
-                                Assert.True(prop.IsRequired);
-                            }
-                        }
-                    }
-                }
-            };
-
-            ClassWithRequiredCustomAttributes obj = new();
-            string json = """{"NonRequired":null,"RequiredA":null,"RequiredB":null}""";
-            Assert.Equal(json, JsonSerializer.Serialize(obj, options));
-
-            var deserialized = JsonSerializer.Deserialize<ClassWithRequiredCustomAttributes>(json, options);
-            Assert.Null(deserialized.NonRequired);
-            Assert.Null(deserialized.RequiredA);
-            Assert.Null(deserialized.RequiredB);
-
-            JsonException exception = Assert.Throws<JsonException>(() => JsonSerializer.Deserialize<ClassWithRequiredCustomAttributes>("{}", options));
-            Assert.Contains(nameof(ClassWithRequiredCustomAttributes.NonRequired), exception.Message);
-            Assert.Contains(nameof(ClassWithRequiredCustomAttributes.RequiredA), exception.Message);
-            Assert.Contains(nameof(ClassWithRequiredCustomAttributes.RequiredB), exception.Message);
-
-            exception = Assert.Throws<JsonException>(() => JsonSerializer.Deserialize<ClassWithRequiredCustomAttributes>("""{"NonRequired":"foo"}""", options));
-            Assert.DoesNotContain(nameof(ClassWithRequiredCustomAttributes.NonRequired), exception.Message);
-            Assert.Contains(nameof(ClassWithRequiredCustomAttributes.RequiredA), exception.Message);
-            Assert.Contains(nameof(ClassWithRequiredCustomAttributes.RequiredB), exception.Message);
-
-            exception = Assert.Throws<JsonException>(() => JsonSerializer.Deserialize<ClassWithRequiredCustomAttributes>("""{"RequiredA":null}""", options));
-            Assert.Contains(nameof(ClassWithRequiredCustomAttributes.NonRequired), exception.Message);
-            Assert.DoesNotContain(nameof(ClassWithRequiredCustomAttributes.RequiredA), exception.Message);
-            Assert.Contains(nameof(ClassWithRequiredCustomAttributes.RequiredB), exception.Message);
-
-            exception = Assert.Throws<JsonException>(() => JsonSerializer.Deserialize<ClassWithRequiredCustomAttributes>("""{"NonRequired":null,"RequiredA":null}""", options));
-            Assert.DoesNotContain(nameof(ClassWithRequiredCustomAttributes.NonRequired), exception.Message);
-            Assert.DoesNotContain(nameof(ClassWithRequiredCustomAttributes.RequiredA), exception.Message);
-            Assert.Contains(nameof(ClassWithRequiredCustomAttributes.RequiredB), exception.Message);
-        }
-
-        [Fact]
-        public static void RequiredExtensionDataPropertyThrows()
-        {
-            ClassWithRequiredCustomAttributeAndDataExtensionProperty obj = new();
-            Assert.Throws<InvalidOperationException>(() => JsonSerializer.Serialize(obj));
-            Assert.Throws<InvalidOperationException>(() => JsonSerializer.Deserialize<ClassWithRequiredCustomAttributeAndDataExtensionProperty>("""{"Data":{}}"""));
-        }
-
-        [Fact]
-        public static void RequiredExtensionDataPropertyCanBeFixedToNotBeRequiredWithResolver()
-        {
-            JsonSerializerOptions options = new()
-            {
-                TypeInfoResolver = new DefaultJsonTypeInfoResolver()
-                {
-                    Modifiers =
-                    {
-                        ti =>
-                        {
-                            if (ti.Type == typeof(ClassWithRequiredCustomAttributeAndDataExtensionProperty))
-                            {
-                                JsonPropertyInfo prop = ti.Properties[0];
-                                Assert.Equal(nameof(ClassWithRequiredCustomAttributeAndDataExtensionProperty.Data), prop.Name);
-                                Assert.True(prop.IsRequired);
-                                prop.IsRequired = false;
-                                Assert.False(prop.IsRequired);
-                            }
-                        }
-                    }
-                }
-            };
-
-            ClassWithRequiredCustomAttributeAndDataExtensionProperty obj = new()
-            {
-                Data = new Dictionary<string, object>()
-                {
-                    ["foo"] = "bar"
-                }
-            };
-
-            string json = """{"foo":"bar"}""";
-            Assert.Equal(json, JsonSerializer.Serialize<ClassWithRequiredCustomAttributeAndDataExtensionProperty>(obj, options));
-            var deserialized = JsonSerializer.Deserialize<ClassWithRequiredCustomAttributeAndDataExtensionProperty>(json, options);
-            Assert.NotNull(deserialized);
-            Assert.NotNull(deserialized.Data);
-            Assert.Equal("bar", ((JsonElement)deserialized.Data["foo"]).GetString());
-        }
-
-        [Fact]
-        public static void RequiredExtensionDataPropertyCanBeFixedToNotBeExtensionDataWithResolver()
-        {
-            JsonSerializerOptions options = new()
-            {
-                TypeInfoResolver = new DefaultJsonTypeInfoResolver()
-                {
-                    Modifiers =
-                    {
-                        ti =>
-                        {
-                            if (ti.Type == typeof(ClassWithRequiredCustomAttributeAndDataExtensionProperty))
-                            {
-                                JsonPropertyInfo prop = ti.Properties[0];
-                                Assert.Equal(nameof(ClassWithRequiredCustomAttributeAndDataExtensionProperty.Data), prop.Name);
-                                Assert.True(prop.IsExtensionData);
-                                prop.IsExtensionData = false;
-                                Assert.False(prop.IsExtensionData);
-                            }
-                        }
-                    }
-                }
-            };
-
-            ClassWithRequiredCustomAttributeAndDataExtensionProperty obj = new()
-            {
-                Data = new Dictionary<string, object>()
-                {
-                    ["foo"] = "bar"
-                }
-            };
-
-            string json = """{"Data":{"foo":"bar"}}""";
-            Assert.Equal(json, JsonSerializer.Serialize<ClassWithRequiredCustomAttributeAndDataExtensionProperty>(obj, options));
-            var deserialized = JsonSerializer.Deserialize<ClassWithRequiredCustomAttributeAndDataExtensionProperty>(json, options);
-            Assert.NotNull(deserialized);
-            Assert.NotNull(deserialized.Data);
-            Assert.Equal("bar", ((JsonElement)deserialized.Data["foo"]).GetString());
-            Assert.Throws<JsonException>(() => JsonSerializer.Deserialize<ClassWithRequiredCustomAttributeAndDataExtensionProperty>("{}", options));
-        }
-
-        [Fact]
-        public static void RequiredReadOnlyPropertyThrows()
-        {
-            ClassWithRequiredCustomAttributeAndReadOnlyProperty obj = new();
-            Assert.Throws<InvalidOperationException>(() => JsonSerializer.Serialize(obj));
-            Assert.Throws<InvalidOperationException>(() => JsonSerializer.Deserialize<ClassWithRequiredCustomAttributeAndReadOnlyProperty>("""{"Data":{}}"""));
-        }
-
-        [Fact]
-        public static void RequiredReadOnlyPropertyCanBeFixedToNotBeRequiredWithResolver()
-        {
-            JsonSerializerOptions options = new()
-            {
-                TypeInfoResolver = new DefaultJsonTypeInfoResolver()
-                {
-                    Modifiers =
-                    {
-                        ti =>
-                        {
-                            if (ti.Type == typeof(ClassWithRequiredCustomAttributeAndReadOnlyProperty))
-                            {
-                                JsonPropertyInfo prop = ti.Properties[0];
-                                Assert.Equal(nameof(ClassWithRequiredCustomAttributeAndReadOnlyProperty.SomeProperty), prop.Name);
-                                Assert.True(prop.IsRequired);
-                                Assert.Null(prop.Set);
-                                prop.IsRequired = false;
-                                Assert.False(prop.IsRequired);
-                            }
-                        }
-                    }
-                }
-            };
-
-            ClassWithRequiredCustomAttributeAndReadOnlyProperty obj = new();
-
-            string json = """{"SomeProperty":"SomePropertyInitialValue"}""";
-            Assert.Equal(json, JsonSerializer.Serialize<ClassWithRequiredCustomAttributeAndReadOnlyProperty>(obj, options));
-
-            json = """{"SomeProperty":"SomeOtherValue"}""";
-            var deserialized = JsonSerializer.Deserialize<ClassWithRequiredCustomAttributeAndReadOnlyProperty>(json, options);
-            Assert.NotNull(deserialized);
-            Assert.Equal("SomePropertyInitialValue", deserialized.SomeProperty);
-
-            json = "{}";
-            deserialized = JsonSerializer.Deserialize<ClassWithRequiredCustomAttributeAndReadOnlyProperty>(json, options);
-            Assert.NotNull(deserialized);
-            Assert.Equal("SomePropertyInitialValue", deserialized.SomeProperty);
-        }
-
-        [Fact]
-        public static void RequiredReadOnlyPropertyCanBeFixedToBeWritableWithResolver()
-        {
-            JsonSerializerOptions options = new()
-            {
-                TypeInfoResolver = new DefaultJsonTypeInfoResolver()
-                {
-                    Modifiers =
-                    {
-                        ti =>
-                        {
-                            if (ti.Type == typeof(ClassWithRequiredCustomAttributeAndReadOnlyProperty))
-                            {
-                                JsonPropertyInfo prop = ti.Properties[0];
-                                Assert.Equal(nameof(ClassWithRequiredCustomAttributeAndReadOnlyProperty.SomeProperty), prop.Name);
-                                Assert.True(prop.IsRequired);
-                                Assert.Null(prop.Set);
-                                prop.Set = (obj, value) => ((ClassWithRequiredCustomAttributeAndReadOnlyProperty)obj).SetSomeProperty((string)value);
-                                Assert.NotNull(prop.Set);
-                            }
-                        }
-                    }
-                }
-            };
-
-            ClassWithRequiredCustomAttributeAndReadOnlyProperty obj = new();
-
-            string json = """{"SomeProperty":"SomePropertyInitialValue"}""";
-            Assert.Equal(json, JsonSerializer.Serialize<ClassWithRequiredCustomAttributeAndReadOnlyProperty>(obj, options));
-
-            json = """{"SomeProperty":"SomeOtherValue"}""";
-            var deserialized = JsonSerializer.Deserialize<ClassWithRequiredCustomAttributeAndReadOnlyProperty>(json, options);
-            Assert.NotNull(deserialized);
-            Assert.Equal("SomeOtherValue", deserialized.SomeProperty);
-
-            json = "{}";
-            JsonException exception = Assert.Throws<JsonException>(() => JsonSerializer.Deserialize<ClassWithRequiredCustomAttributeAndReadOnlyProperty>(json, options));
-            Assert.Contains("SomeProperty", exception.Message);
-        }
-
         public class ClassWithNullableProperty
         {
             private string? _value;
@@ -2030,38 +1698,6 @@ namespace System.Text.Json.Serialization.Tests
             {
                 get => _value;
                 set => _value = value ?? "NULL";
-            }
-        }
-
-        public class ClassWithRequiredCustomAttributes
-        {
-            [JsonPropertyOrder(0)]
-            public string NonRequired { get; set; }
-
-            [JsonPropertyOrder(1)]
-            [JsonRequired]
-            public string RequiredA { get; set; }
-
-            [JsonPropertyOrder(2)]
-            [JsonRequired]
-            public string RequiredB { get; set; }
-        }
-
-        public class ClassWithRequiredCustomAttributeAndDataExtensionProperty
-        {
-            [JsonRequired]
-            [JsonExtensionData]
-            public Dictionary<string, object>? Data { get; set; }
-        }
-
-        public class ClassWithRequiredCustomAttributeAndReadOnlyProperty
-        {
-            [JsonRequired]
-            public string SomeProperty { get; private set; } = "SomePropertyInitialValue";
-
-            public void SetSomeProperty(string value)
-            {
-                SomeProperty = value;
             }
         }
     }
